requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langchain==0.3.7
langchain-openai==0.2.3
openai==1.51.2
python-docx==1.1.2
